update to unsigned int

+++ src/core/ed_config.h
@@ -26,15 +26,15 @@ enum class DiagonalizationMethod;
  * @brief Core diagonalization parameters
  */
 struct DiagonalizationConfig {
-    int num_eigenvalues = 1;
-    int max_iterations = 10000;
+    uint64_t num_eigenvalues = 1;
+    uint64_t max_iterations = 10000;
     double tolerance = 1e-10;
     bool compute_eigenvectors = false;
     
     // Method-specific
     double shift = 0.0;           // For shift-invert
-    int block_size = 4;          // For block methods
-    int max_subspace = 100;       // For Davidson
+    uint64_t block_size = 4;          // For block methods
+    uint64_t max_subspace = 100;       // For Davidson
     double target_lower = 0.0;    // For Chebyshev filtered (lower energy bound)
     double target_upper = 0.0;    // For Chebyshev filtered (upper energy bound)
 };
@@ -43,31 +43,31 @@ struct DiagonalizationConfig {
  * @brief Thermal calculation parameters
  */
 struct ThermalConfig {
-    int num_samples = 1;
+    uint64_t num_samples = 1;
     double temp_min = 1e-3;
     double temp_max = 20.0;
-    int num_temp_bins = 100;
+    uint64_t num_temp_bins = 100;
     
     // TPQ-specific
-    int num_order = 100;          // Order for cTPQ
-    int num_measure_freq = 100;   // Measurement frequency
+    uint64_t num_order = 100;          // Order for cTPQ
+    uint64_t num_measure_freq = 100;   // Measurement frequency
     double delta_tau = 1e-2;      // Time step for cTPQ
     double large_value = 1e5;     // Large value for mTPQ
     
     // FTLM-specific
-    int ftlm_krylov_dim = 100;    // Krylov subspace dimension per sample
+    uint64_t ftlm_krylov_dim = 100;    // Krylov subspace dimension per sample
     bool ftlm_full_reorth = false; // Use full reorthogonalization
-    int ftlm_reorth_freq = 10;    // Reorthogonalization frequency
-    unsigned int ftlm_seed = 0;   // Random seed (0 = auto)
+    uint64_t ftlm_reorth_freq = 10;    // Reorthogonalization frequency
+    uint64_t ftlm_seed = 0;   // Random seed (0 = auto)
     bool ftlm_store_samples = false; // Store per-sample intermediate data
     bool ftlm_error_bars = true;  // Compute error bars
     
     // LTLM-specific
-    int ltlm_krylov_dim = 200;    // Krylov subspace dimension for excitations
-    int ltlm_ground_krylov = 100; // Krylov dimension for finding ground state
+    uint64_t ltlm_krylov_dim = 200;    // Krylov subspace dimension for excitations
+    uint64_t ltlm_ground_krylov = 100; // Krylov dimension for finding ground state
     bool ltlm_full_reorth = false; // Use full reorthogonalization
-    int ltlm_reorth_freq = 10;    // Reorthogonalization frequency
-    unsigned int ltlm_seed = 0;   // Random seed (0 = auto)
+    uint64_t ltlm_reorth_freq = 10;    // Reorthogonalization frequency
+    uint64_t ltlm_seed = 0;   // Random seed (0 = auto)
     bool ltlm_store_data = false; // Store intermediate data
     
     // Hybrid Thermal (LTLM+FTLM standalone method)
@@ -86,7 +86,7 @@ struct ObservableConfig {
     // Spectral functions
     double omega_min = -10.0;
     double omega_max = 10.0;
-    int num_points = 1000;
+    uint64_t num_points = 1000;
     
     // Time evolution
     double t_end = 50.0;
@@ -103,20 +103,20 @@ struct ObservableConfig {
 struct DynamicalResponseConfig {
     bool calculate = false;
     bool thermal_average = false;     // If true, compute thermal-averaged response
-    int num_random_states = 20;       // Number of random states for thermal averaging
-    int krylov_dim = 100;             // Krylov subspace dimension
+    uint64_t num_random_states = 20;       // Number of random states for thermal averaging
+    uint64_t krylov_dim = 100;             // Krylov subspace dimension
     double omega_min = -10.0;         // Minimum frequency
     double omega_max = 10.0;          // Maximum frequency
-    int num_omega_points = 1000;      // Number of frequency points
+    uint64_t num_omega_points = 1000;      // Number of frequency points
     double broadening = 0.1;          // Lorentzian broadening parameter
     double temp_min = 0.01;           // Minimum temperature (for temperature scan)
     double temp_max = 10.0;           // Maximum temperature (for temperature scan)
-    int num_temp_bins = 1;            // Number of temperature points (1 = single temperature at temp_min)
+    uint64_t num_temp_bins = 1;            // Number of temperature points (1 = single temperature at temp_min)
     bool compute_correlation = false; // Compute two-operator correlation
     std::string operator_file = "";   // File containing operator to probe
     std::string operator2_file = "";  // Second operator file (for correlation)
     std::string output_prefix = "dynamical_response";
-    unsigned int random_seed = 0;     // Random seed (0 = auto)
+    uint64_t random_seed = 0;     // Random seed (0 = auto)
 };
 
 /**
@@ -124,31 +124,31 @@ struct DynamicalResponseConfig {
  */
 struct StaticResponseConfig {
     bool calculate = false;
-    int num_random_states = 20;       // Number of random states for thermal averaging
-    int krylov_dim = 100;             // Krylov subspace dimension
+    uint64_t num_random_states = 20;       // Number of random states for thermal averaging
+    uint64_t krylov_dim = 100;             // Krylov subspace dimension
     double temp_min = 0.01;           // Minimum temperature
     double temp_max = 10.0;           // Maximum temperature
-    int num_temp_points = 100;        // Number of temperature points
+    uint64_t num_temp_points = 100;        // Number of temperature points
     bool compute_susceptibility = true; // Compute susceptibility dO/dT
     bool compute_correlation = false; // Compute two-operator correlation
     bool single_operator_mode = false; // Compute single-operator ⟨O⟩ instead of ⟨O†O⟩
     std::string operator_file = "";   // File containing operator(s) to probe
     std::string operator2_file = "";  // Second operator file (for correlation)
     std::string output_prefix = "static_response";
-    unsigned int random_seed = 0;     // Random seed (0 = auto)
+    uint64_t random_seed = 0;     // Random seed (0 = auto)
 };
 
 /**
  * @brief System/lattice parameters
  */
 struct SystemConfig {
-    int num_sites = 0;
+    uint64_t num_sites = 0;
     float spin_length = 0.5;
-    int sublattice_size = 1;
+    uint64_t sublattice_size = 1;
     
     // Fixed Sz mode
     bool use_fixed_sz = false;
-    int n_up = -1;  // Number of up spins (-1 = not set, will use num_sites/2)
+    uint64_t n_up = -1;  // Number of up spins (-1 = not set, will use num_sites/2)
     
     std::string hamiltonian_dir = "";
     std::string interaction_file = "InterAll.dat";
@@ -161,8 +161,8 @@ struct SystemConfig {
 struct ArpackConfig {
     bool verbose = false;
     std::string which = "SR";  // For Hermitian matrices: SR=Smallest Real (ground state), LR=Largest Real
-    int ncv = -1;
-    int max_restarts = 2;
+    uint64_t ncv = -1;
+    uint64_t max_restarts = 2;
     double ncv_growth = 1.5;
     bool auto_enlarge_ncv = true;
     bool two_phase_refine = true;
@@ -174,7 +174,7 @@ struct ArpackConfig {
     bool adaptive_inner_tol = true;
     double inner_tol_factor = 1e-2;
     double inner_tol_min = 1e-14;
-    int inner_max_iter = 300;
+    uint64_t inner_max_iter = 300;
 };
 
 /**
@@ -219,12 +219,12 @@ public:
     // ========== Builder Methods ==========
     
     // Diagonalization
-    EDConfig& eigenvalues(int n) { diag.num_eigenvalues = n; return *this; }
-    EDConfig& iterations(int n) { diag.max_iterations = n; return *this; }
+    EDConfig& eigenvalues(uint64_t n) { diag.num_eigenvalues = n; return *this; }
+    EDConfig& iterations(uint64_t n) { diag.max_iterations = n; return *this; }
     EDConfig& tolerance(double t) { diag.tolerance = t; return *this; }
     EDConfig& eigenvectors(bool b = true) { diag.compute_eigenvectors = b; return *this; }
     EDConfig& shift(double s) { diag.shift = s; return *this; }
-    EDConfig& blockSize(int b) { diag.block_size = b; return *this; }
+    EDConfig& blockSize(uint64_t b) { diag.block_size = b; return *this; }
     EDConfig& targetRange(double lower, double upper) { 
         diag.target_lower = lower; 
         diag.target_upper = upper; 
@@ -232,27 +232,27 @@ public:
     }
     
     // Thermal
-    EDConfig& samples(int n) { thermal.num_samples = n; return *this; }
+    EDConfig& samples(uint64_t n) { thermal.num_samples = n; return *this; }
     EDConfig& tempRange(double min, double max) { 
         thermal.temp_min = min; 
         thermal.temp_max = max; 
         return *this; 
     }
-    EDConfig& tempBins(int n) { thermal.num_temp_bins = n; return *this; }
+    EDConfig& tempBins(uint64_t n) { thermal.num_temp_bins = n; return *this; }
     
     // Observable
     EDConfig& calcObservables(bool b = true) { observable.calculate = b; return *this; }
     EDConfig& measureSpin(bool b = true) { observable.measure_spin = b; return *this; }
     
     // System
-    EDConfig& numSites(int n) { system.num_sites = n; return *this; }
+    EDConfig& numSites(uint64_t n) { system.num_sites = n; return *this; }
     EDConfig& spinLength(float s) { system.spin_length = s; return *this; }
     EDConfig& hamiltonianDir(const std::string& dir) { 
         system.hamiltonian_dir = dir; 
         return *this; 
     }
     EDConfig& fixedSz(bool use = true) { system.use_fixed_sz = use; return *this; }
-    EDConfig& numUp(int n) { system.n_up = n; return *this; }
+    EDConfig& numUp(uint64_t n) { system.n_up = n; return *this; }
     
     // Workflow
     EDConfig& standard(bool b = true) { workflow.run_standard = b; return *this; }
@@ -273,7 +273,7 @@ public:
     /**
      * @brief Parse command line arguments
      */
-    static EDConfig fromCommandLine(int argc, char* argv[]);
+    static EDConfig fromCommandLine(uint64_t argc, char* argv[]);
     
     /**
      * @brief Merge configurations (command line overrides file)
@@ -291,7 +291,7 @@ public:
     void save(const std::string& filename) const;
     
     /**
-     * @brief Print configuration summary
+     * @brief Pruint64_t configuration summary
      */
     void print(std::ostream& out = std::cout) const;
     
@@ -329,7 +329,7 @@ public:
         const std::string& category = "General"
     );
     
-    bool parse(int argc, char* argv[]);
+    bool parse(uint64_t argc, char* argv[]);
     
     std::optional<std::string> get(const std::string& name) const;
     bool has(const std::string& name) const;
